Move "DrawOfferPopup" component to "popups" directory in "features"
